Delete ui file and put code in main.rs
