+++ src/main.rs
@@ -1,3 +1,5 @@
+use gtk::prelude::*;
+use relm4::prelude::*;
 use std::convert::TryInto;
 use std::env;
 use std::fs;
@@ -24,6 +26,77 @@ use crate::util::elf_structure::ELFIdentification;
 use crate::util::elf_structure::FileInfoELF;
 
 use crate::util::macho_structure::MachOHeader;
+
+struct App {
+    pub counter: u8,
+}
+
+#[derive(Debug)]
+enum Msg {
+    Increment,
+    Decrement,
+}
+
+#[relm4::component]
+impl SimpleComponent for App {
+    type Init = u8;
+    type Input = Msg;
+    type Output = ();
+
+    view! {
+        gtk::Window {
+            set_title: Some("Simple app"),
+            set_default_size: (300, 100),
+
+            gtk::Box {
+                set_orientation: gtk::Orientation::Vertical,
+                set_spacing: 5,
+                set_margin_all: 5,
+
+                gtk::Button {
+                    set_label: "Increment",
+                    connect_clicked => Msg::Increment,
+                },
+
+                gtk::Button {
+                    set_label: "Decrement",
+                    connect_clicked => Msg::Decrement,
+                },
+
+                gtk::Label {
+                    #[watch]
+                    set_label: &format!("Counter: {}", model.counter),
+                    set_margin_all: 5,
+                }
+            }
+        }
+    }
+
+    // Initialize the component.
+    fn init(
+        counter: Self::Init,
+        root: Self::Root,
+        sender: ComponentSender<Self>,
+    ) -> ComponentParts<Self> {
+        let model = App { counter };
+
+        // Insert the code generation of the view! macro here
+        let widgets = view_output!();
+
+        ComponentParts { model, widgets }
+    }
+
+    fn update(&mut self, msg: Self::Input, _sender: ComponentSender<Self>) {
+        match msg {
+            Msg::Increment => {
+                self.counter = self.counter.wrapping_add(1);
+            }
+            Msg::Decrement => {
+                self.counter = self.counter.wrapping_sub(1);
+            }
+        }
+    }
+}
 
 // context
 struct Ctx {
@@ -450,9 +523,11 @@ fn get_file_data(file_signature: &str, bytes: &[u8]) {
 // ===========================================================================
 
 fn main() {
+    let app = RelmApp::new("relm4.example.simple");
+    app.run::<App>(0);
     //help();
-    let context: Ctx = get_arguments();
-    let bytecode = read_file(&context.filename);
-    let sign = get_sign(&bytecode);
-    get_file_data(&sign, &bytecode);
+    //let context: Ctx = get_arguments();
+    //let bytecode = read_file(&context.filename);
+    //let sign = get_sign(&bytecode);
+    //get_file_data(&sign, &bytecode);
 }
